@@ -1,17 +1,19 @@
-import { RouteHandlerMethod } from "fastify";
+import { AccountType } from "@prisma/client";
+import { randomUUID } from "crypto";
 
 import { authService } from "./auth.services";
+import { LoginSchema } from "./auth.schemas";
+import { RouteHandlerMethodTypebox } from "../../types";
 
-export const loginHandler: RouteHandlerMethod = async (request, reply) => {
+export const loginHandler: RouteHandlerMethodTypebox<
+    typeof LoginSchema
+> = async (request, reply) => {
     const sub = request.user.sub;
 
     // search for the user with that sub
     let user = await request.prisma.user.findFirst({
-        where: {
-            subs: {
-                has: sub,
-            },
-        },
+        where: { subs: { has: sub } },
+        include: { account: { include: { plan: true } } },
     });
 
     if (!user) {
@@ -28,6 +30,7 @@ export const loginHandler: RouteHandlerMethod = async (request, reply) => {
         if (!email || !name) {
             // id token don't have email and name, does not allow to log in
             return reply.code(401).send({
+                statusCode: 401,
                 error: "Unauthorized",
                 message:
                     "User did not provide access to email and name properties",
@@ -36,35 +39,73 @@ export const loginHandler: RouteHandlerMethod = async (request, reply) => {
 
         // search again for user using email
         user = await request.prisma.user.findUnique({
-            where: {
-                email,
-            },
+            where: { email },
+            include: { account: { include: { plan: true } } },
         });
 
         if (user) {
             // found user by email, just add the sub to that
+            // also update the name, in case it changed
             user = await request.prisma.user.update({
-                where: {
-                    email,
-                },
-                data: {
-                    name, // also update the name, in case it changed
-                    subs: {
-                        push: sub,
-                    },
-                },
+                where: { email },
+                data: { name, subs: { push: sub } },
+                include: { account: { include: { plan: true } } },
             });
         } else {
-            // could not find user, create it
+            // could not find user, create it now, along with user account
+            const id = randomUUID();
+
+            // search for default plan of user account, throw error if not configured
+            const plan = await request.prisma.plan.findFirstOrThrow({
+                where: {
+                    accountTypes: { has: AccountType.USER },
+                    default: true,
+                },
+            });
+
+            // create stripe customer
+            const customerId = await request.server.billing.createCustomer({
+                email,
+                name,
+                description: name,
+                metadata: { external_id: id },
+            });
+
+            // create user entity
             user = await request.prisma.user.create({
                 data: {
+                    id,
                     email,
                     name,
                     subs: [sub],
+                    account: {
+                        create: {
+                            id,
+                            type: AccountType.USER,
+                            plan: { connect: { id: plan.id } },
+                        },
+                    },
+                    billingCustomerId: customerId,
+                },
+                include: {
+                    account: { include: { plan: true } },
                 },
             });
         }
     }
 
-    return reply.code(200).send(user);
+    // handle stripe subscription
+    if (!user.account.billingSubscriptionId) {
+        // create a checkout session using the configured plan as the stripe price object
+        const checkoutUrl = await request.server.billing.createCheckoutUrl({
+            account: user.account,
+            email: user.email,
+        });
+
+        return reply
+            .code(200)
+            .send({ email: user.email, subscription: { url: checkoutUrl } });
+    } else {
+        return reply.code(200).send({ email: user.email });
+    }
 };
