@@ -1,16 +1,15 @@
 import { FastifyPluginAsyncTypebox } from "@fastify/type-provider-typebox";
 import { FastifyTypebox } from "../../types";
-import { loginHandler } from "./auth.handlers";
-import { LoginSchema } from "./auth.schemas";
+import { listHandler } from "./plans.handlers";
+import { ListPlansSchema } from "./plans.schemas";
 
 const routes: FastifyPluginAsyncTypebox = async (server: FastifyTypebox) => {
-    server.post(
-        "/login",
+    server.get(
+        "/",
         {
-            schema: LoginSchema,
-            preValidation: server.authenticate,
+            schema: ListPlansSchema,
         },
-        loginHandler
+        listHandler
     );
 };
 
