@@ -1,7 +1,10 @@
 import { PrismaClient } from "@prisma/client";
-import { File, Suite } from "vitest";
+import { File, Suite, beforeEach } from "vitest";
+import { Stripe } from "stripe";
 import buildServer from "../../src/server";
 import { createDatabase } from "./database";
+import { FastifyContext } from "../types";
+import { TestStripeBillingManager } from "../billing";
 
 const buildServerHook = async (ctx: Suite | File) => {
     // create database for test
@@ -13,9 +16,18 @@ const buildServerHook = async (ctx: Suite | File) => {
     });
     await prisma.$connect();
 
+    // build an additional stripe client (server will create its own)
+    const stripe = new Stripe(process.env.STRIPE_SECRET_KEY!, {
+        apiVersion: "2022-11-15",
+    });
+
+    // instantiate a test billing manager
+    const billing = new TestStripeBillingManager(stripe);
+
     // build the server
     const server = buildServer({
         prisma,
+        billing,
         logger: false,
     });
 
@@ -23,10 +35,26 @@ const buildServerHook = async (ctx: Suite | File) => {
     ctx.meta = {
         prisma,
         server,
+        stripe,
+        billing,
     };
 
+    beforeEach<FastifyContext>(async (ctx) => {
+        // copy suite context to test context
+        ctx.prisma = prisma;
+        ctx.server = server;
+        ctx.stripe = stripe;
+        ctx.billing = billing;
+    });
+
     // return clean up function, to disconnect from database
-    return async () => await prisma.$disconnect();
+    return async () => {
+        // delete all billing data created for this suite
+        await billing.teardown();
+
+        // disconnect from database
+        await prisma.$disconnect();
+    };
 };
 
 export default buildServerHook;
